Add `load()` calls for messages fonts
Aside from maybe only Chrome, simply adding preload links for the font
files is insufficient to get the initial layout to render with the fonts
since the FontFace still only loads when the font is first used, i.e.
when the messages get added to the DOM.

Also remove preload link for Noto Sans JP, because the file apparently
varies based on platform.