Standardize the backgrounds of all pages
1. Sets default background color to navbar's red

Mainly an Easter egg for Android Chrome users where the OS bar will take
the colour of the background.

2. Uses the same gradient effect for the other pages

The gradient overlay used on the Timeline page started from a light gray
to white instead of a white to white that was previously used as I found
it less harsh on the eyes.