Use Miku's Steel Rose tag as start of Timeline
Source: https://x.com/Mikururun/status/1836300917227831656/photo/1

Needed something at the top of the regular sized Timeline because the
vertical centering of Events leaves a little bit of line just orphaned
not doing anything.

Uses CSS Grids to help centre and ensure the <img> is constrained to a
certain fraction of the Timeline width. Currently set to 1/5th.