Install minijinja-cli via download
The existing method of installing `minijinja-cli` via `cargo install` is
very slow as it has to install all the Rust deps from scratch every
single run. This change switches to just installing the precompiled
binaries using the script mentioned in the minijinja repo.

https://github.com/mitsuhiko/minijinja

Instead of taking over a minute to finish the `Build` job when `cargo
install` is used this decreases the time to under 20 seconds.