Attempt a less jank grid for the markers
Instead of having a 0.5rem grid column with the markers on top we make
the column fit the markers and have the "line" drawn the same way its
done for the `baetag`.

By having the columns contain the markers we can make use of
`column-gap` to control the spacing between the markers and the other
boxes. Also lets us get rid of the jank transforms to centre the
markers.